+++ CHANGELOG.md
@@ -13,11 +13,21 @@ This is the official first release that is considered stable.
 
 ### Noteworthy additions and changes
 
+- [DZNPY-0030](https://github.com/mikeftrict/dznpy/issues/30): In Advanced Shell generate std::ref
+  port functors instead of copying entire ports for MTS ports. The old mechanism 'worked' for Dezyne
+  versions until and including 2.17. But for 2.18+ not. With std-reffing each in-event, as what
+  Dezyne does, it is ensured that this particular code is compatible with all Dezyne versions (
+  including 2.18) so far.
+    - Also added in this issue/merge-request is a DummyToaster and DummyExclusiveToaster
+      implementation model that comes with an Advanced Shell with only 1 provides port and (because
+      it is a stub) no requires ports.
+
 - [DZNPY-0029](https://github.com/mikeftrict/dznpy/issues/29): Generate explicit constructor port
   initialization to be closer with the 2.18/2.19 code generation.
     - Also, the `CommentBlock` class is renamed to `Comment` (replacing the old one). It uses the
       the new TextBlock features.
     - Solved a bunch of PyLint issues.
+
 - [DZNPY-0028](https://github.com/mikeftrict/dznpy/issues/28): Moved `TextBlock` and its cohesively
   related constructs to a (new) module `text_gen`. Also integrate bullet list numbering
   functionality into `TextBlock` itself as part of indentation. Lastly, it adds a header content